renamed pysixtracklib in sixtracklib
