+++ python/.gitignore
@@ -4,5 +4,5 @@ dist/
 build/
 *.egg-info/
 
-pysixtracklib/__pycache__/
+sixtracklib/__pycache__/
