+++ python/CMakeLists.txt
@@ -4,8 +4,8 @@ if( NOT PYTHONINTERP_FOUND )
     message( FATAL_ERROR "Python interpreter not found -> stopping" )
 endif()
 
-add_subdirectory( pysixtracklib )
-add_subdirectory( pysixtracklib_test )
+add_subdirectory( sixtracklib )
+add_subdirectory( sixtracklib_test )
 
 set( PYTHON_SIXTRACKLIB_SHARED_LIB      ${PY_SIXTRACKLIB_SHARED_LIB}  )
 set( PYTHON_SIXTRACKLIB_TEST_SHARED_LIB ${PY_SIXTRACKLIB_TEST_SHARED_LIB} )
@@ -19,14 +19,14 @@ configure_file( ${SETUP_PY_IN_PATH} ${SETUP_PY_OUT_PATH} @ONLY )
 
 if( SIXTRACKL_INSTALL_PYSIXTRACKLIB )
     set( PYSIXTRACKLIB_INSTALL_STAMP
-         "${CMAKE_BINARY_DIR}/python/pysixtracklib/pysixtracklib.stamp" )
+         "${CMAKE_BINARY_DIR}/python/sixtracklib/sixtracklib.stamp" )
 
     add_custom_command( OUTPUT ${PYSIXTRACKLIB_INSTALL_STAMP}
                         COMMAND ${PYTHON_EXECUTABLE} ${SETUP_PY_OUT_PATH} build
                         COMMAND ${CMAKE_COMMAND} -E touch ${PYSIXTRACKLIB_INSTALL_STAMP}
     )
 
-    add_custom_target( pysixtracklib ALL DEPENDS
+    add_custom_target( sixtracklib ALL DEPENDS
                        ${PYSIXTRACKLIB_INSTALL_STAMP} )
 
     install( CODE
