+++ python/setup.py.template
@@ -1,18 +1,18 @@
 import setuptools
 
 setuptools.setup(
-    name='pysixtracklib',
+    name='sixtracklib',
     version='@SIXTRACKL_VERSION_STR@',
     description='Python bindings for SixTrackLib (6D-Tracking Code)',
     author='SixTrack',
     author_email='sixtrack.contact@cern.ch',
     url='https://github.com/SixTrack/sixtracklib',
-    packages=['pysixtracklib', 'pysixtracklib_test' ],
-    package_dir= { 'pysixtracklib': 'pysixtracklib',
-                   'pysixtracklib_test': 'pysixtracklib_test', },
+    packages=['sixtracklib', 'sixtracklib_test' ],
+    package_dir= { 'sixtracklib': 'sixtracklib',
+                   'sixtracklib_test': 'sixtracklib_test', },
     package_data = {
-        'pysixtracklib': [ 'pysixtracklib/@PYTHON_SIXTRACKLIB_SHARED_LIB@' ],
-        'pysixtracklib_test': [ 'pysixtracklib_test/@PYTHON_SIXTRACKLIB_TEST_SHARED_LIB@' ]
+        'sixtracklib': [ 'sixtracklib/@PYTHON_SIXTRACKLIB_SHARED_LIB@' ],
+        'sixtracklib_test': [ 'sixtracklib_test/@PYTHON_SIXTRACKLIB_TEST_SHARED_LIB@' ]
     },
     install_requires=['numpy','cobjects>=0.0.4','pysixtrack>=0.0.0','sixtracktools>=0.0.2']
 )
