+++ python/sixtracklib/config.py.template
