+++ python/sixtracklib/particles.py
