+++ python/sixtracklib/stcommon.py
